#527 Fix by polling even mappings for which control is disabled
This has a nice effect: It makes the time when the button is pressed
relevant and lets subsequent bank/modifier changes not have any influence
on the merits of "Fire after timeout", "Single press" etc. That's
actually what we want. A modifier doesn't need to be pressed the whole
time, just in the beginning. Nice!